add tests

+++ src/main/java/com/countries/model/BordersAsia.java
@@ -1,6 +1,6 @@
 package com.countries.model;
 
-public record BordersAsia(String country, Long borders) implements Comparable<BordersAsia> {
+public record BordersAsia(String country, Integer borders) implements Comparable<BordersAsia> {
 
     @Override
     public int compareTo(BordersAsia o) {
